refactor(pytorch): organize CUDA step metadata kernels
